requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
email-validator==2.2.0
gunicorn==21.2.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
